requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
SQLAlchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0.post1
